+++ lapis/src/main/kotlin/org/genspectrum/lapis/response/SiloResponse.kt
@@ -1,17 +1,10 @@
 package org.genspectrum.lapis.response
 
 import com.fasterxml.jackson.core.JsonGenerator
-import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.core.type.TypeReference
-import com.fasterxml.jackson.databind.DeserializationContext
-import com.fasterxml.jackson.databind.JsonDeserializer
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.JsonSerializer
 import com.fasterxml.jackson.databind.SerializerProvider
-import com.fasterxml.jackson.databind.node.ObjectNode
 import io.swagger.v3.oas.annotations.media.Schema
-import org.genspectrum.lapis.config.DatabaseConfig
-import org.genspectrum.lapis.util.UNALIGNED_PREFIX
 import org.springframework.boot.jackson.JsonComponent
 
 const val COUNT_PROPERTY = "count"
