+++ lapis/src/main/kotlin/org/genspectrum/lapis/silo/SiloQuery.kt
@@ -4,105 +4,24 @@ import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.core.JsonGenerator
-import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.JsonSerializer
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
-import com.fasterxml.jackson.databind.node.BooleanNode
-import com.fasterxml.jackson.databind.node.DoubleNode
-import com.fasterxml.jackson.databind.node.FloatNode
-import com.fasterxml.jackson.databind.node.IntNode
-import com.fasterxml.jackson.databind.node.LongNode
-import com.fasterxml.jackson.databind.node.NullNode
-import com.fasterxml.jackson.databind.node.TextNode
-import org.apache.arrow.vector.BigIntVector
-import org.apache.arrow.vector.BitVector
-import org.apache.arrow.vector.Float4Vector
-import org.apache.arrow.vector.Float8Vector
-import org.apache.arrow.vector.IntVector
-import org.apache.arrow.vector.VarCharVector
-import org.apache.arrow.vector.VectorSchemaRoot
 import org.genspectrum.lapis.request.OrderByField
 import org.genspectrum.lapis.request.OrderBySpec
 import org.genspectrum.lapis.response.AggregationData
-import org.genspectrum.lapis.response.COUNT_PROPERTY
 import org.genspectrum.lapis.response.DetailsData
 import org.genspectrum.lapis.response.InsertionData
 import org.genspectrum.lapis.response.MostCommonAncestorData
 import org.genspectrum.lapis.response.MutationData
 import org.genspectrum.lapis.response.PhyloSubtreeData
 import org.genspectrum.lapis.response.SequenceData
-import org.genspectrum.lapis.util.UNALIGNED_PREFIX
 import java.time.LocalDate
 
 data class SiloQuery<ResponseType>(
     val action: SiloAction<ResponseType>,
     val filterExpression: SiloFilterExpression,
 )
-
-/** Converts one row at `rowIndex` from an Arrow [VectorSchemaRoot] to a typed Kotlin object. */
-typealias ArrowRowConverter<T> = (root: VectorSchemaRoot, rowIndex: Int) -> T
-
-private fun VectorSchemaRoot.getString(
-    name: String,
-    rowIndex: Int,
-): String? {
-    val vector = getVector(name) as? VarCharVector ?: return null
-    if (vector.isNull(rowIndex)) {
-        return null
-    }
-    return String(vector.get(rowIndex))
-}
-
-private fun VectorSchemaRoot.getInt(
-    name: String,
-    rowIndex: Int,
-): Int? {
-    val vector = getVector(name) as? IntVector ?: return null
-    if (vector.isNull(rowIndex)) {
-        return null
-    }
-    return vector.get(rowIndex)
-}
-
-private fun VectorSchemaRoot.getLong(
-    name: String,
-    rowIndex: Int,
-): Long? {
-    val vector = getVector(name) as? BigIntVector ?: return null
-    if (vector.isNull(rowIndex)) {
-        return null
-    }
-    return vector.get(rowIndex)
-}
-
-private fun VectorSchemaRoot.getDouble(
-    name: String,
-    rowIndex: Int,
-): Double? {
-    val vector = getVector(name) as? Float8Vector ?: return null
-    if (vector.isNull(rowIndex)) {
-        return null
-    }
-    return vector.get(rowIndex)
-}
-
-private fun VectorSchemaRoot.fieldValueAsJsonNode(
-    columnIndex: Int,
-    rowIndex: Int,
-): JsonNode {
-    val vector = fieldVectors[columnIndex]
-    if (vector.isNull(rowIndex)) return NullNode.instance
-    return when (vector) {
-        is VarCharVector -> TextNode(String(vector.get(rowIndex)))
-        is IntVector -> IntNode(vector.get(rowIndex))
-        is BigIntVector -> LongNode(vector.get(rowIndex))
-        is Float8Vector -> DoubleNode(vector.get(rowIndex))
-        is Float4Vector -> FloatNode(vector.get(rowIndex))
-        is BitVector -> BooleanNode.valueOf(vector.get(rowIndex) != 0)
-        else -> TextNode(vector.getObject(rowIndex)?.toString() ?: "")
-    }
-}
 
 interface CommonActionFields {
     val orderByFields: List<OrderByField>
@@ -114,8 +33,8 @@ interface CommonActionFields {
 const val ORDER_BY_RANDOM_FIELD_NAME = "random"
 
 sealed class SiloAction<ResponseType>(
-    @JsonIgnore val cacheable: Boolean,
     @JsonIgnore val arrowConverter: ArrowRowConverter<ResponseType>,
+    @JsonIgnore val cacheable: Boolean,
 ) : CommonActionFields {
     companion object {
         fun aggregated(
@@ -265,18 +184,8 @@ sealed class SiloAction<ResponseType>(
         override val limit: Int? = null,
         override val offset: Int? = null,
     ) : SiloAction<AggregationData>(
+            arrowConverter = AGGREGATION_DATA_ARROW_CONVERTER,
             cacheable = true,
-            arrowConverter = { root, rowIndex ->
-                val count = root.getLong(COUNT_PROPERTY, rowIndex)?.toInt() ?: 0
-                val fields = root.schema.fields
-                    .filter { it.name != COUNT_PROPERTY }
-                    .mapIndexed { colIdx, field ->
-                        val actualColIdx = root.schema.fields.indexOfFirst { it.name == field.name }
-                        field.name to root.fieldValueAsJsonNode(actualColIdx, rowIndex)
-                    }
-                    .toMap()
-                AggregationData(count, fields)
-            },
         ) {
         val type: String = "Aggregated"
     }
@@ -290,8 +199,8 @@ sealed class SiloAction<ResponseType>(
         override val offset: Int? = null,
         val fields: List<String> = emptyList(),
     ) : SiloAction<MutationData>(
-            cacheable = true,
             arrowConverter = MUTATION_DATA_ARROW_CONVERTER,
+            cacheable = true,
         ) {
         val type: String = "Mutations"
     }
@@ -305,8 +214,8 @@ sealed class SiloAction<ResponseType>(
         override val offset: Int? = null,
         val fields: List<String> = emptyList(),
     ) : SiloAction<MutationData>(
-            cacheable = true,
             arrowConverter = MUTATION_DATA_ARROW_CONVERTER,
+            cacheable = true,
         ) {
         val type: String = "AminoAcidMutations"
     }
@@ -319,14 +228,8 @@ sealed class SiloAction<ResponseType>(
         override val limit: Int? = null,
         override val offset: Int? = null,
     ) : SiloAction<DetailsData>(
+            arrowConverter = DETAILS_DATA_ARROW_CONVERTER,
             cacheable = false,
-            arrowConverter = { root, rowIndex ->
-                DetailsData(
-                    root.schema.fields.mapIndexed { colIdx, field ->
-                        field.name to root.fieldValueAsJsonNode(colIdx, rowIndex)
-                    }.toMap(),
-                )
-            },
         ) {
         val type: String = "Details"
     }
@@ -338,8 +241,8 @@ sealed class SiloAction<ResponseType>(
         override val limit: Int? = null,
         override val offset: Int? = null,
     ) : SiloAction<InsertionData>(
-            cacheable = true,
             arrowConverter = INSERTION_DATA_ARROW_CONVERTER,
+            cacheable = true,
         ) {
         val type: String = "Insertions"
     }
@@ -353,14 +256,8 @@ sealed class SiloAction<ResponseType>(
         override val offset: Int? = null,
         override val randomize: RandomizeConfig? = null,
     ) : SiloAction<MostCommonAncestorData>(
+            arrowConverter = MOST_COMMON_ANCESTOR_DATA_ARROW_CONVERTER,
             cacheable = true,
-            arrowConverter = { root, rowIndex ->
-                MostCommonAncestorData(
-                    mrcaNode = root.getString("mrcaNode", rowIndex),
-                    missingNodeCount = root.getInt("missingNodeCount", rowIndex) ?: 0,
-                    missingFromTree = root.getString("missingFromTree", rowIndex),
-                )
-            },
         ) {
         val type: String = "MostRecentCommonAncestor"
     }
@@ -374,14 +271,8 @@ sealed class SiloAction<ResponseType>(
         override val offset: Int? = null,
         override val randomize: RandomizeConfig? = null,
     ) : SiloAction<PhyloSubtreeData>(
+            arrowConverter = PHYLO_SUBTREE_DATA_ARROW_CONVERTER,
             cacheable = true,
-            arrowConverter = { root, rowIndex ->
-                PhyloSubtreeData(
-                    subtreeNewick = root.getString("subtreeNewick", rowIndex) ?: "",
-                    missingNodeCount = root.getInt("missingNodeCount", rowIndex) ?: 0,
-                    missingFromTree = root.getString("missingFromTree", rowIndex),
-                )
-            },
         ) {
         val type: String = "PhyloSubtree"
     }
@@ -393,8 +284,8 @@ sealed class SiloAction<ResponseType>(
         override val limit: Int? = null,
         override val offset: Int? = null,
     ) : SiloAction<InsertionData>(
-            cacheable = true,
             arrowConverter = INSERTION_DATA_ARROW_CONVERTER,
+            cacheable = true,
         ) {
         val type: String = "AminoAcidInsertions"
     }
@@ -409,40 +300,9 @@ sealed class SiloAction<ResponseType>(
         val sequenceNames: List<String>,
         val additionalFields: List<String> = emptyList(),
     ) : SiloAction<SequenceData>(
+            arrowConverter = SEQUENCE_DATA_ARROW_CONVERTER,
             cacheable = false,
-            arrowConverter = { root, rowIndex ->
-                SequenceData(
-                    root.schema.fields
-                        .mapIndexed { colIdx, field ->
-                            field.name.removePrefix(UNALIGNED_PREFIX) to root.fieldValueAsJsonNode(colIdx, rowIndex)
-                        }
-                        .toMap(),
-                )
-            },
         )
-}
-
-private val MUTATION_DATA_ARROW_CONVERTER: ArrowRowConverter<MutationData> = { root, rowIndex ->
-    MutationData(
-        mutation = root.getString("mutation", rowIndex),
-        count = root.getInt("count", rowIndex),
-        proportion = root.getDouble("proportion", rowIndex),
-        sequenceName = root.getString("sequenceName", rowIndex),
-        mutationFrom = root.getString("mutationFrom", rowIndex),
-        mutationTo = root.getString("mutationTo", rowIndex),
-        position = root.getInt("position", rowIndex),
-        coverage = root.getInt("coverage", rowIndex),
-    )
-}
-
-private val INSERTION_DATA_ARROW_CONVERTER: ArrowRowConverter<InsertionData> = { root, rowIndex ->
-    InsertionData(
-        count = root.getInt("count", rowIndex) ?: 0,
-        insertion = root.getString("insertion", rowIndex) ?: "",
-        insertedSymbols = root.getString("insertedSymbols", rowIndex) ?: "",
-        position = root.getInt("position", rowIndex) ?: 0,
-        sequenceName = root.getString("sequenceName", rowIndex) ?: "",
-    )
 }
 
 sealed class SiloFilterExpression(
